refactor, add docs, update Makefile
